archived old plaid implementation and refactored new implementation and test
